net_smtp: Add basic delivery failure emails.
Adds a minimally functional delivery failure
autoresponse email mechanism for queued emails
that we fail to deliver successfully.
These emails are not being generated according
to the specifications currently, but a response
does go out and the original email is attached.
Future improvements can generate the email with
the relevant MIME parts and headers.