event.c: Add event bus API.
Adds an internal event bus that can be used
to decouple production of events with actions
that consume them.

Currently, there are not very many of these
(an existing instance of such a consumer is
moved into the new mod_events module), but
there will likely be more in the future.

This decoupling means that consumers can be
modified in modules without having to restart
the BBS (consumers are more likely to change
than producers are).

The event types themselves are integer
values, which is not as extensible as using
strings but provides better performance.