net_imap: Add LSUB command.
Adds the LSUB command, which is just the same
as LIST in this implementation. We essentially
force clients to subscribe to all available
folders, so the SUBSCRIBE and UNSUBSCRIBE commands
don't do anything functional. LSUB is needed at
minimum since clients expect to be able to obtain
a list of subscribed folders in normal operation.