utils: Fix copy_file_range portability issues.
Apparently, copy_file_range is not sufficiently portable,
as the syscall isn't even available on one system under
testing, causing all copies to fail. Therefore, we now
wrap this in a BBS function that will attempt to use
copy_file_range if the syscall is available, but fall back
to using sendfile if not.