IRC: Allow DMing between IRC and other networks.
Allows users on the native IRC server to have
direct messages with users on other IRC networks,
as well as Discord users, through mod_relay_irc
and mod_discord respectively.

Users on the "other side" must prefix their messages
with the name of the user on the native IRC server
when messaging the service account on that network.