net_smtp: Fix relaying reliability.
Fixes issues with bbs_expect not reading an entire
line of data and thus thinking the expected output
wasn't received when it really would be if it were
properly detected.

Also includes a slew of other miscellaneous fixes
and improvements. We now add the Received header
on all messages received by the server, including
those submitted by message submission agents, to
include the most information, preserving the original
sender's privacy if directed. I know of at least one
ISP whose SMTP server does this.

I encountered a strange bug with TCP behaving very
wonky if multiple instances of the BBS were running,
so we now default to not reusing ports by default,
and doing so only if necessary, to avoid masking this
problem unnecessarily.

To avoid hanging if connect() fails to connect,
we also set a timeout for connect so that it will
return early if the connection cannot be set up.

Finally, in module.c, don't try to load modules that
we know will fail to load anyways due to missing
dependencies.

A known issue at this time seems to be that emails
cannot be sent in realtime directly, but they are
successfully sent when queued. The issue is with
connect() itself, which does not return successfully
in the realtime case. This needs to be further investigated.