net_smtp: Fix realtime message delivery.
For some reason, the immediate sending of
messages using the smtp data structure
no longer works. As hinted in commit
70a1249ba7fed1e8884860847f672e33a30731d6,
the TCP connection just hangs with connect(),
regardless of whether it is blocking or
nonblocking. This is possibly due to some
recent regression (though not sure what),
but we may as well take advantage of this bug
to add an option to allow for messages to be
sent asynchronously (i.e. not during the SMTP session itself).

This improves responsiveness for SMTP clients.
Obviously, there are pros and cons of doing it
this way, and these are noted in the sample config.

Normally, this is a configurable option; until
the relevant bug is fixed, however, it is forcibly
enabled through a macro definition in the code,
so that messages can always be sent immediately.