net_smtp: Add option to notify about delayed deliveries.
Adds an option that allows users who send messages that
are not delivered initially to be notified each time
message delivery is retried. This is in addition to (and
separate from) delivery failure notifications.

I don't know if any other mail servers do this, but this
can be a nice feature to have to know if your message
wasn't delivered initially, since these days, if it's
not delivered immediately, there is a high probability
that delivery will fail, and this provides more immediate
feedback than waiting for the bounce message, potentially
much later.