mod_auth_mysql: Warn about TTY/TDD password compatibility.
Users may not use passwords containing lowercase letters
(or many symbols) if they wish to be able to log in from
a TTY/TDD. This is because these devices do not support
those characters (indeed, 5-bit Baudot code can't encode
them).

There isn't really a good solution or workaround to this
limitation. Case-insensitive password comparisons can't
be done without storing the password, which we don't. So
just warn the user about this so they can use a compatible
password if desired.

This patch also contains some (unrelated) refactoring of
some SMTP/NNTP code.