mod_auth_static: Add static config backed user auth.
This adds an authentication module that is backed by
a static config, as opposed to mod_auth_mysql which
is backed by a database.

This is not intended for production usage. It anticipates
usage in black box testing where it would be useful to
mock authentication in this manner. Therefore, this
module should not be loaded, except when running such tests.