net_nntp: Fix past POSTs prepending to subsequent ones.
The post body is not reset after an article submission is
processed, which can lead to previous articles, whether
accepted or not, being prepended to future articles submitted
on the same connection. The post data is now explicitly
reset to fix this.

This issue was caught during the development of the currently
in-progress black box testing framework.