mod_mailscript: Add MailScript rule engine.
This adds a highly flexible, dynamic scripting engine
called "MailScript" to the SMTP server that can be used
to perform arbitrary operations on incoming and outgoing
email messages using an "if this, then that" type formula.

Currently, this is intended for system administrators and
there is no direct way for BBS users to create their own
rules, as this can have security implications. However,
common user operations such as filtering, forwarding,
spam filtering, etc. can all be accomplished through
the use of MailScript rules. A custom transpiler to generate
MailScript rules using such user input could be used to do this,
but that is not in scope currently.

A global .rules file in the root maildir directory applies
to all mailboxes, and each user may also have a .rules file
in his or her own maildir root.

Some black box tests are included to test the functionality
included in this commit.

The SMTP server also now counts Received headers in order
to detect mail loops.