linkedlists: Refactor remove element operations.
Removing a specific element from a linked list is
a common operation throughout the BBS.

This adds the RWLIST_REMOVE_BY_FIELD and
RWLIST_WRLOCK_REMOVE_BY_FIELD macros to linkedlists.h
to streamline operations that remove a single element
if a member of the struct in the list matches a particular
value.

Additionally, some operations are refactored to use the
preexisting RWLIST_REMOVE API.

These changes simplify a large number of linked list element
removal operations throughout the tree.