test_imap: Add IMAP tests.
Adds black box tests for IMAP, as well as adds
to the existing POP3 tests.

As part of this change, the option to run the BBS
under valgrind in the test suite is also included.
This can help catch issues that a normal execution
of the test suite would not.

Several bug fixes to several modules uncovered as
part of testing are included in this commit.