tests: Add various black box tests.
Adds tests for a number of things, namely:
- net_finger
- net_ftp
- net_gopher
- net_http
- Menu functionality

Some existing tests are also enhanced.
Some of these tests are not super comprehensive,
but I wanted to get something in to touch everything
possible, which is better than nothing. The HTTP and FTP
tests in particular could probably use more rigorous
client expectations.