net_smtp: Properly fetch and iterate MX records.
The original MX lookup was not fully RFC
compliant, as it would merely use the first
MX result that was returned, even though
the results were not sorted, so this was not
even necessarily the MX server with the lowest
number (highest precedence) priority.

We now fetch all the MX records and sort them,
and retry subsequent MX servers if initial ones
fail.

We do not randomly choose between MX servers with
the same priority number, as the RFC directs, but
otherwise this is now significantly more RFC-compliant,
and should not result in delivery failures due to
not iterating over all available MX results.