alloc.c: Simplify memory allocation handling.
This adds a thin abstraction layer for memory allocation,
similar to the one that exists for file descriptors in fd.c.
Although this does not track memory, it does provide common
functions that will log an error if an allocation fails,
which means the codebase no longer needs to be littered
with warnings wherever allocations appear for off-nominal paths.
These have been removed to simplify allocation within allocating
code.

A test is added that randomly fails allocations and ensures nothing
goes too awry. However, this is not run by default when all tests
are run currently since it does not pass consistently yet. However,
a good number of issues fixed in initial runs of this test (leaks
on off-nominal paths, etc.) have been fixed.