memset: Fix optimized out buffer sanitization.
Most existing uses of memset to zero out buffers
containing sensitive data are likely to be optimized
away since their values are not used subsequently.
This replaces these calls to memset with calls to
a wrapper around explicit_bzero to ensure this zeroing
out does not get optimized away by the compiler.