net_ftp: Add CDUP and MDTM commands.
Adds the CDUP and MDTM commands, since these
are both attempted by CoreFTP. Supporting CDUP
means that clients won't have to fallback to
CWD if it's not supported.

CoreFTP interestingly attempts to use MDTM
to store the modification timestamp of a file
rather than using it to retrieve the modification
timestamp. This appears to be noncompliant
with the relevant specifications, so the added
code is not used by CoreFTP after all, but, well,
it's there now.