door_irc: Allow 1:1 client connections to native IRC server.
Previously, there was no easy way to use the BBS's IRC client
(door_irc) to connect to the BBS's IRC server (net_irc) in
a sensible way. The mechanism to do this entailed defining
a persistent client in door_irc.conf using the credentials of
some BBS user (such as the sysop, or more ideally, a service account)
which would relay messages from door_irc to the IRC server.
However, this makes all messages from door_irc appears as if
they came from a single IRC user, and the overhead to setting this
up is not insignificant.

This is now simplified greatly by adding the option to have door_irc
make a 1:1 connection ad hoc, for a single client, to any arbitrary
IRC server. Consequently, it is now possible in menus.conf to have
a menu option that will connect to a channel on the native IRC server.
The semantics of this are discussed in the door_irc sample config.
Because a separate connection can be made for each door_irc session,
clients accessing IRC in this way appear just as normal clients, and
the same user can even be logged in from multiple nodes (which is typically
not permitted via IRC).