Makefiles: Add dependencies on header files.
Previously, header files were not make targets
in any way, which meant that changes to header
files would not in and of themselves cause any
source files including that header file to be
recompiled. Normally, this is not an issue for
users. However, for developers, this can be a
minor irritant as if ABI is broken (e.g. inserting
a member into a struct in the middle), then
running "make" could result in object code
referencing members at different addresses, generally
leading to a SEGV fairly quickly. The workaround has been
to run "make clean" whenever header files are modified,
but this is obviously far from ideal.

As a quality of life improvement for developers,
the Makefiles now include the header file dependencies
for every source file. These are automatically generated
using the -MMD option the first time source files are
compiled. Any changes to any non-library header files
will ensure any dependent source files are subsequently recompiled.

Currently, the .d dependency files are only built on the
first compile (thus the dummy dependency to satisfy these
targets initially). At this point, these targets exist as files
(since they are not PHONY), so make will see these as changed
and recompile all the source files a second time. However,
after the BBS has been compiled twice, only future header file
changes will result in recompiling the relevant files, as expected.