door_irc: Don't reprompt for password for loopback IRC.
Commit 2bd47dcb305fe2ac05d8e60d3aa01fe2ae951a7a added the
ability to relatively seamlessly establish connections to
the locally running IRC server from within the BBS. However,
it required prompting the user again for his/her password,
since this is needed for authentication.

This introduces a workaround by using one-time use tokens
that can be used for authenticating a user in lieu of a
password. This allows the token to be automatically generated
on the client side and verified on the server side. As a result,
when used in this mode, the user is immediately connected
to the channel, resulting in a much more seamless experience.