socket.c: Fix bind port reuse regression.
Commit 70a1249ba7fed1e8884860847f672e33a30731d6 changed
bbs_make_tcp_socket to not reuse the port by default,
since this could mask things that shouldn't have been
happening (such as multiple BBS processes running).
Instead, we now emit a warning if binding without reuse
failed, and then bind with reuse to force the bind to
succeed, which at least makes it more obvious that there
is potentially a problem (but sometimes, this can be
ignored if the socket is merely in the CLOSE_WAIT state).

The aforementioned commit happened to flip a few lines
of code in the process of making the relevant changes.
However, it seems that flipping the order of filling the
sockaddr_in struct and calling setsockopt results in
the reuse not actually working. There was code to retry
bind up to 100 times for this reason, and eventually it
would succeed, but only because the socket was no longer
in CLOSE_WAIT at that point. If SO_REUSEPORT was really
having an effect, it should succeed immediately.

The order of these two sections is now flipped back to how
it was originally, and the port reuse not works as expected
in both the in use and not in use cases, though I'm still
not exactly clear on why this is. Because reuse now works
properly, the retry/sleep code is removed and we only attempt
to bind a maximum of 2 times.