branches: Add extremely minor branch locality hints.
Sprinkles a few likely() and unlikely() macros throughout
the BBS core. These don't directly impact branch prediciton,
but they can result in more optimal code layout when
branches are encountered to favor the likely (or disfavor
the unlikely) branch.

These macros have been added in spots where, in normal execution
of the BBS, a branch will be taken 0% or 100% of the time.
This is most useful in alloc.c and logger.c, which are called
extremely frequently throughout the codebase. The performance
benefit is probably negligible since branch predictors are pretty
good now, but this won't hurt anything.