transfer.c: Add user home directories.
This allows users on the BBS to have a home directory
that is currently accessible only to them for all operations.
In the future, perhaps the user will be able to control
access at a more granular level, e.g. if certain users wish
to allow others read access to certain files in their home
directory.