net_pop3: Fix type mismatch.
Commit 5e523ba20bbacf66a1fff38bbb30465ad04f9fea changed
the POP3 deletion tracking to use a char (1 byte)
instead of an int (8 bytes) so that blocks were properly
allocated in 1-byte chunks to fix some memory issues.
However, the type was still referenced as int in some places,
which caused this to not work properly and one of the unit
tests to fail. This is now fixed by using sizeof(char)
instead of sizeof(int) everywhere needed.

A few memory leaks in other unit tests are also fixed.

Additionally, a black box test module is added that simply
runs the unit tests, and succeeds if 100% of them pass.
This is added since there's no elegant way of running
the unit tests in the CI so this is an easy way to do that.