net_smtp: Allow submitted messages to be relayed.
Adds a RELAY MailScript filtering action that will
relay the received message through another SMTP server
(typically a message submission agent).

One use case for this is this grants clients the ability
to send messages through multiple SMTP servers simply
by configuring a single SMTP server in their client.
By adjusting the From address, the client can control
which SMTP server is used to send the message. Effectively,
this offloads configuration from the client to the server.
Note that for IMAP clients, the sent message is still stored
in the BBS's Sent folder.

This can also be useful for clients that wish to mask their
IP address. For example, Google includes the client's
original IP address in the Received header that it tacks on.
Relaying messages through the BBS (if at a different IP address)
can help preserve user privacy.