net_imap: Fix broken FETCH responses.
There were a few problems with FETCH responses,
alluded to by my IMAP client failing to fetch
messages and confirmed by dovecot's fantastic
IMAPTester client.

The format for the INTERNALDATE was wrong, and
this has now been fixed.

Critically, there was a missing space between
the length literal and any previous content,
which led to parsing failures. This is now fixed.

Additionally, the size was not consistent with
the body length; this is made consistent, although
the fix in this patch may not necessarily be the
best place to do this and should be revisited.