net_imap: Fix FETCH HEADER.FIELDS regression.
Commit 2e93a7262ad5aff0264796780dc17ca588688047
accidentally switched from using strstr to strncmp
for matching headers, which is wrong because the match
variable contains a list of headers, not just a single header.

That part is fixed, while retaining the essence of what
the previous commit was trying to fix, ensuring that
headers don't match too leniently. To do this efficiently,
we duplicate the list of headers and modify it in such a way
that we can correctly use strstr on the result.