tests: Use readline_data for reliable expect.
io_relay in will use strstr to check if a string
being expected appears in the buffer of data read
from the BBS console's STDOUT, but this is not reliable
since read can return an arbitrary number of bytes,
and chunks may not correspond cleanly to lines.
Messages that were expected being split across multiple
reads therefore led to occasional test execution failures.
The expected string was present, but we were not reading
it in a robust manner.

This refactors some code so that the test suite can link
to the readline_data code and use the bbs_fd_readline
functions to more robustly read output data. This should
address the spurious test failures.