utils: Add file existence wrapper.
Checking for file existence is a common
operation throughout the tree. Previously,
we were mostly using eaccess to do this,
even though using stat may be faster.
This replaces many of these eaccess calls
with abstract calls to the new wrapper function,
to better convey semantics and allow us to
change the implementation if needed.