user.c: Cache username/user ID translations.
Converting between user IDs and usernames is a
common operation throughout the BBS. Currently,
this is an extremely inefficient operation, since
with the limited existing APIs for talking to the
auth provider, we are loading all the users from
the database into structures, retrieving the user
we want from that, and then picking just the field
we want from that. This operation has a lot of overhead
and is mostly wasteful.

It would still be ideal to extend the auth provider
APIs so we can retrieve certain info about a specific
user, rather than having to retrieve all users.
However, even more optimally for these two operations
would just be to cache this information, since we do
not expect it to change. This is now done and results
in significantly faster translation times, since we
don't need to talk to the database and we don't need
allocate lots of memory that will never be used.