net_imap: Fix SEARCH and FETCH command issues.
Fixes a few more issues exposed by Dovecot ImapTest.

First, this adds support for using sequence numbers
in the SEARCH command input, which as specified in
the RFC, are sequence numbers that merely act as a
filter in and of themselves, where present.

Secondly, if there are no matches, we were not sending
any untagged response. We now send an empty untagged
response if appropriate.

Thirdly, for FETCH ENVELOPE, if a name contained quotes,
the quotes would be sent without being escaped. We now
strip the quotes to avoid this (escaping would also work).

If the SEARCH query contained parentheses to group
multiple search keys, these were not parsed and handled
properly. We now apply order of operations to parentheses.

Occasionally date-related SEARCH queries would be wrong,
and this is fixed by initializing the tm1 and tm2 structs
to 0. I don't entirely like this solution but it was
not apparent how they were being used uninitialized.

Finally, there was an issue with ACL parsing where if
an ACL file existed but there was no match for the user,
then the ACL was effectively 0. This is now fixed to use
the default for the user/mailbox combo instead in that case.