thread.c: Simplify thread cancel and kill.
Calling pthread_cancel followed by pthread_kill
is a common sequence throughout the BBS for thread
cleanup, so this is refactored into a common
BBS function that emits warnings automatically
as appropriate.

Locking in pty.c is also adjusted in order to
fix a deadlock that could occur if the node was
locked waiting for the PTY master thread to exit,
but the PTY master thread was trying acquire the
node lock at the top of the loop.