system.c: Finish isoexec container execution.
Code was added back in commit 6b640e8ea6e9712e58c5b662afd41d4c021ea504
to support the use of namespaces to provide isolation
when executing external programs. However, enhancement
of this functionality stalled for a while as debugging
was done to determine an ideal way of integrating this.

Ultimately, initial approaches to simply create a root filesystem
for the container on the fly by symlinking existing directories
as needed failed, and here we instead expect the sysop to provide
his or her own root filesystem. For testing, I have used a
Debian 11 rootfs that was around 130 MB in a tarball. The sysop
can install utilities to be available on the BBS beforehand.

This change effectively allows containers to be run within
the BBS. Docker uses namespaces and cgroups to run containers,
for example; here, we only use namespaces, but the sysop
could manually use cgroups to control resource usage as well.

Currently, the rootfs is persistent and shared between all users.
It is therefore more directly suited for providing isolation
from the main system, not isolated between users. In theory,
each user could have a separate rootfs, and this would provide
isolation and privacy between users. However, I'm not sure many
sysops would want N copies of a 130 MB filesystem on their machine,
particularly since most of it is the same. Further work will need
to be done to only duplicate necessary files, but providing a
separate rootfs environment for each user would be a worthwhile enhancement.

This change now allows sysops to do more interestings things like
provide shell access on the BBS, without granting access to the root
namespace/filesystem itself. This technique is more secure than mechanisms
such as chroot, which have known security holes that allow users to
break out of chroot jails. However, sysops should still understand
the full security implications of this functionality before utilizing it.