net_imap, net_smtp: Add BURL SMTP; fix IMAP, SMTP, and TLS bugs.
This adds support for RFC 4468 BURL SMTP. Normally, sent
email messages must be uploaded twice: once to the SMTP
message submission agent for sending the message, and then
to the IMAP server to store it in the "Sent" folder. This is
wasteful and duplicative, but this is still standard today.

BURL SMTP allows a client to first upload the message to the
IMAP server and then send the IMAP URL to the SMTP server
for submission, allowing the SMTP server to retrieve the
message, rather than requiring a second upload. Since in simple
cases, the SMTP and IMAP server may be one and the same (as for the BBS),
this makes a lot of sense.

Unfortunately, BURL SMTP is not widely supported by servers, and
is even less widely supported by clients. Trojita is the only
known IMAP client in existence to support this feature that is
currently known about. This functionality has thus been verified
to work with that implementation; even though we do not need
IMAP GENURLAUTH because of the trust relationship that allows
us to pull messages directly from the SMTP server module, we have
a minimal stub that makes the Trojita client happy. Note that generic
BURL support is not implemented here; only BURL support specific for
the native IMAP server; the SMTP server itself does not support URLFETCH.
In most cases, this should be fine since most people's SMTP and
IMAP servers belong to the same entity.

Finally, this also adds the ability for users to send mail as other
users (including non-user mailboxes), since users may want to send
mail using the address of a mailbox that is shared with them.
Additionally, the "anyone" and "authenticated" targets are properly
recognized for IMAP ACLs, as specified in RFC 4314, and + and -
are properly handled for the SETACL command.

Other minor enhancements in this change:
* Use RFC 5530 IMAP responses wherever possible

This also fixes a few major bugs that are not directly to BURL SMTP:
* tls.c did not set up the TLS I/O relay thread if the TLS server was disabled,
  which led to an abort when trying to set up a TLS client. Since these use
  the TLS I/O thread as well, we now always spawn the thread, since TLS clients
  do not require any configuration to use.
* When an IP address is used as the SMTP hostname, it must be enclosed in []
  square brackets as per the RFC. Failing to do this caused other SMTP servers
  to reject messages from servers with an IP hostname; this is now fixed.
* The IMAP server was not actually properly sorting messages to determine the right
  sequence numbers to use. In particular, alphasort was used, which results in a
  consistent ordering over the same set of messages, but not necessarily the
  right ordering. This is wrong as the RFC states that sequence numbers MUST
  be in order of UID. Erroneous behavior can ensue if a mailbox has messages whose
  filenames are not also in order of UID, which can happen when messages are
  moved or copied between mailboxes. Case in point, Thunderbird clients behave
  very erratically when the wrong ordering is used; messages will spontaneously
  disappear and reappear in every other ordering of a mailbox.
  To fix this, instead of using alphasort, we now use a comparison function
  for scandir that results in an ordering of filenames by UID, resulting in
  consistent behavior for clients.
* EXPUNGE responses MUST NOT be sent when no command is in progress, or during
  a FETCH, STORE, SEARCH, or SORT command, according to the RFCs. Previously,
  we would send responses to any clients selecting the same mailbox immediately.
  We now delay these responses using a pipe until the appropriate time.