net_imap: Enhance SEARCH capabilities.
Adds support for ESEARCH responses as defined in
RFC4466 and RFC4731 (MIN, MAX, COUNT, ALL).

Adds support for the SEARCHRES extension, whereby
a client can save the result of a search and reference
it in the future using the '$' symbol (RFC5182).

Adds support for RFC5267 ESORT.

Fixes the UID SEARCH key to refer to a sequence set rather
than just a specific UID.

Also fixes unilaterial responses to always send them immediately
if a client is idling and always buffer them if not. We now
also send unilateral FETCH responses to other clients
when message flags are updated.

An unrelated hotfix also included here avoids closing
file descriptors in net_ssh if they were never set.