net_imap: Add further IMAP extensions, fix keyword portability.
Adds support for:
* RFC 7889 APPENDLIMIT
* RFC 6851 MOVE
* RFC 5032 WITHIN (OLDER, YOUNGER)

This includes a few tweaks to prepare for future
RFC 7162 CONDSTORE support, but that functionality is
not actually present here, and will be added later.

A major oversight addressed in this commit is that, as
flags and keywords are stored in the filename, and the
mappings are not necessarily consistent between maildirs,
the flag/keyword portion of the filename MAY change when
a message is copied or moved to a different mailbox.
In particular, keywords do not have a constant mapping
throughout the IMAP server, or even between a single user's folders.
Thus, blindly moving a message without adjusting these could
result in keywords changing. Proper translation is now done
when these operations occur.