net_imap: Add RFC 7162 QRESYNC support.
As part of this change, HIGHESTMODSEQ is now stored persistently
on disk in a binary file, since we need to keep track of
UIDs and MODSEQs for expunged files.

The per-file directory scan callbacks for EXPUNGE and CLOSE are
removed and scanned directly in a single function so that we can
do processing for every single expunged message efficiently at
the end. For example, we can update HIGHESTMODSEQ at the end all
at once, rather than after each individual deletion.

Unfortunately, CONDSTORE is not super widely supported and QRESYNC
is even less widely supported. Tests have been added to the test
suite for these changes, but limited functional testing has been
done because Thunderbird-based clients are still very buggy in
their CONDSTORE support, and don't support QRESYNC at all.