mod_events: Improve IP autoban functionality.
This significantly improves the autoban functionality
in mod_events, which greatly reduces the number of
false positives that were previously caught.
For example, Thunderbird's mail server auto detection
would happily trip the previous logic and immediately
ban the client. At the same time, the ban detection
was not very robust either.

This now stores more state per IP address so that it
can more accurately determine when a client is being
malicious and deserves to be blocked or not.

Most private IP address ranges are now also ignored.

Additionally, in net_ssh, we now explicitly generate
a short session event if a connection is made and exits
without a PTY ever having been established. This does
not happen on legitimate SSH connections, so these can
be treated as garbage traffic.