mail: Add modular email mailer system.
Originally, the BBS was hardcoded to use the
sendmail utility in mail.c to send emails,
regardless of the destination. This was fine
then as there was no SMTP module at the time,
but now that the BBS is itself a mail server,
it no longer makes sense to prefer sendmail
for sending email when we can do it directly,
with much less overhead and increased efficiency.

This change introduces the concept of a "mailer",
with a configurable priority, to allow different
modules to accept responsibility for delivering
email. The SMTP module is now the preferred mailer,
and sending mail using sendmail is moved into its
own module (mod_sendmail), and still registers as
a backup. The BBS core therefore no longer needs
to be concerned without how the email itself is
delivered.

The most important impact here is that other parts
of the BBS can send emails to users, including
local users directly, without being concerned about
how that is being done. Future changes will build
on top of this improvement.