net_imap: Add virtual mailboxes (IMAP client proxy to remote server).
This adds an extremely basic IMAP proxy feature. This can be used to view
a mailbox on another server in the folder hierarchy of a local IMAP account.
This allows a client to connect to multiple IMAP servers using a single
set of credentials. It functions as a proxy in the sense that the connection
is mostly proxied passthrough, and could be used for example if one leg or
another is using TLS or not, a client supports different capabilities, etc.

This is still incomplete; notable, there is no APPEND support currently,
and COPY and MOVE operations that cross server boundaries will fail.
In theory, these could be made to work if they transparently transformed
into an APPEND when possible, assuming the remote server supports UIDPLUS.

An important regression fix bundled in this patch is a partial revert of
commit 385a7fc32bc0f7537206fdb42ce2b072c3b0676b. A variable increment
was erroneously removed in tls.c due to cppcheck saying the statement
had no effect. In fact, this is not true, and this broke TLS connections.

In addition to this, SSL_write returning SSL_WANT_WRITE was not handled,
leading to failed writes, and this now addressed (though perhaps not
in the most equitable way) by retrying.