net_imap: Add SPECIAL-USE and XLIST.
We've already been returning RFC 6154
SPECIAL-USE attributes since the beginning,
but we now also allow filtering LIST
to just return SPECIAL-USE mailboxes, and
thus now return SPECIAL-USE as a capability.

Additionally, the nonstandard and deprecated
XLIST has been added. This is because many
clients, in practice, support the older XLIST
but not the newer RFC 6154 usage. For example,
Microsoft Outlook seems to create its own Sent
and Trash folders if XLIST is not available.
The response format is virtually identical to
LIST and LSUB, so there's no overhead to supporting
this for the clients that need it.