net_sieve, mod_sieve: Add Sieve and ManageSieve protocols.
This adds basic support for RFC 5228 Sieve filtering and
the RFC 5804 ManageSieve protocol. Sieve is a standardized
filtering engine supports by some other mail servers,
so this is intended to supplant MailScript when possible.
However, there are still some things in the BBS implementation
of Sieve that cannot be done with Sieve (e.g. executing
external programs, operating on sent messages, etc.)

mod_sieve is responsible for actually handling the Sieve
filtering itself, and is built on top of libsieve. It
mostly works, but there are a few bugs (including memory-related ones)
in it, it's not very flexible, and it isn't being maintained
(has not seen any updates in over 10 years). For this reason,
it would be best to migrate to a custom Sieve library
that allows us to do the things we want, namely supporting
more recent Sieve extensions and doing more thorough validation
of Sieve scripts. However, the implementation details of this
are all contained to mod_sieve, and neither ManageSieve nor
net_smtp needs to be nominally aware of that.