net_telnet: Add TELNETS support.
This theoretically adds support for TELNETS
(Telnet over TLS), a somewhat esoteric protocol
that allows for encrypted Telnet sessions.
The designated port for TELNETS is TCP 992.

Not tested with an actual TELNETS client,
since I'm not aware of any, but it mostly
works (in the way you would expect) with
openssl's s_client.

Because this requires an explicit TLS relay
for the node, which thus far has not been
needed (only used for non-terminal network
services), a separate read/write file descriptor
are stored on the node. Because the actual node
file descriptor is not used in many places, there
wasn't much code that needed to be updated to refer
to these instead of node->fd directly. And thus
has been added yet another layer of indirection.