net_imap: Add OAuth2 auth proxying, fix IMAP proxying bugs.
This adds the ability for IMAP proxying to servers that
require OAuth2 to be used to log in to an account, e.g.
some Gmail accounts.

To accomplish this, a generic OAuth token provider mechanism
has been added that allows other modules to query the current
OAuth access token for a specific purpose. Some work is required
up front to actually load in the client ID, client secret,
and refresh token, but once we have these, we can transparently
handle OAuth2 on behalf of users. This also allows usage of these
accounts (via the proxy) for programs that do not natively support
OAuth2.

OAuth2 support has also been added for SMTP relaying, for MailScript
rules that RELAY the message to another SMTP message submission agent.
However, there are some security caveats with this implementation,
which are discussed at length in mod_oauth.conf.

OAuth support will NOT be added for POP3, since POP3
doesn't support connecting to other servers; it only provides
access to the local INBOX, so there's no point in adding OAuth support.

Since this was mainly tested with Gmail, some Gmail-related fixes
are also included. In particular, we now properly account for the
remote server using a different hierarchy delimiter than we do,
properly translating back and forth as needed to maintain the
seamless experience for the end user.

Gimap (Google IMAP server) is also quite strict in its command
parsing, and does not tolerate leading or trailing whitespace
between tokens in commands. We are now more careful about ensuring
that extra spaces are not present. This avoids some BAD warnings
that clients will otherwise display to users.

Furthermore, we now keep track of a few more capabilities and
don't allow usage of remote capabilities that we know are not
supported.

Some "extra" (typically) unwanted folders are also omitted from
LIST responses, namely Gmail's All Mail, Starred, and Important
folders, which are not "real" mailboxes, but are virtual mailboxes
that contain mail in other folders. Since this unnecessarily duplicates
messages, these are hidden since IMAP users don't want these.

Also bumps the minor version to 2, which really should have
been done a long time ago, but better late than never.