net_smtp: Add remote IMAP appends, fix possible OAuth security hole.
This adds the ability for the SMTP server to do an IMAP APPEND to a remote IMAP
server, as well as move a message to a local folder. This is useful if users
want the SMTP server to store their sent messages on the IMAP server (either
local or remote) instead of having their mail client do it. This is more bandwidth
efficient since the client only needs to upload the message once, instead of twice;
the server will upload the message twice instead on the client's behalf, and if
the message is being saved locally, no second upload is needed at all.

The previous commit added OAuth relaying support, which was reasonably secure
for IMAP but not for SMTP. We now pass a reference to the SMTP authenticated
user where needed so that SMTP can be secure as well, i.e. we no longer pass
"0" for the user to the OAuth functions.

With these changes, the IMAP/SMTP proxies are reasonably functional and can
be used to thoroughly manage other email accounts through the BBS mail server.

Other fixes in this commit:
* Server Name Indication (SNI) is now sent on TLS client connections.
* Fixes a crash in auth.c if password was NULL.